fix mujoco type conversions
Former-commit-id: a2fad4f895dc30c36fe334559b15ad59d38f5626

+++ rllab/mujoco_py/codegen.rb
@@ -306,7 +306,7 @@ def #{prop[:name]}(self):
 
 @#{prop[:name]}.setter
 def #{prop[:name]}(self, value):
-    val_ptr = np.array(value).ctypes.data_as(POINTER(#{ctype_type}))
+    val_ptr = np.array(value, dtype=np.float64).ctypes.data_as(POINTER(#{ctype_type}))
     memmove(self._wrapped.contents.#{prop[:name]}, val_ptr, #{count} * sizeof(#{ctype_type}))
 }
